EDITOR-#161 fix conflicts
